backend/requirements.txt
flask
stability-sdk
python-dotenv
flask_cors
openai